% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_label.R
\name{clean_label}
\alias{clean_label}
\title{Clean label text}
\usage{
clean_label(expr)
}
\arguments{
\item{expr}{An R expression or call.}
}
\value{
Character string with no whitespace.
}
\description{
Converts an unevaluated R expression to a single, whitespace-free string
suitable for labelling timed operations.
}
\keyword{internal}
