% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emit.R
\name{emit}
\alias{emit}
\title{Emit a timing message}
\usage{
emit(start, end, label, time_unit, console = TRUE, log_file = NULL)
}
\arguments{
\item{start}{A \code{POSIXct} object indicating the start time.}

\item{end}{A \code{POSIXct} object indicating the end time.}

\item{label}{Character string to describe the operation being timed.}

\item{time_unit}{Character; the unit in which to report the duration.}

\item{console}{Logical; whether to print the message to the console.}

\item{log_file}{Optional character path to a file to append the message.}
}
\value{
Invisibly returns the duration of the operation.
}
\description{
Logs or prints the execution duration of an operation.
}
\keyword{internal}
