% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{get_log}
\alias{get_log}
\title{Retrieve a stored timing log}
\usage{
get_log(log)
}
\arguments{
\item{log}{Character. Name of the data frame to load from \code{.pipetime_env}.}
}
\value{
A data frame of timing logs.
}
\description{
Retrieve a stored timing log
}
