% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_log.R
\name{rm_log}
\alias{rm_log}
\title{Remove a stored timing log}
\usage{
rm_log(log)
}
\arguments{
\item{log}{Character. Name of the timing log to delete from \code{.pipetime_env}.}
}
\description{
Remove a stored timing log
}
