% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_pipe.R
\name{time_pipe}
\alias{time_pipe}
\title{Measure execution time in a pipeline}
\usage{
time_pipe(
  .data,
  label = NULL,
  log_file = NULL,
  console = NULL,
  time_unit = NULL
)
}
\arguments{
\item{.data}{The input object to pass through the pipeline.}

\item{label}{Optional. Name for the operation. Defaults to the expression if not provided.}

\item{log_file}{Optional. File to write timing logs.
Can also be set globally via \code{options(pipetime.log_file = "filename.log")}.
Defaults to NULL (no logging).}

\item{console}{Logical. Print messages to the console? Can also be set globally via
\code{options(pipetime.console = TRUE)}. Defaults to TRUE.}

\item{time_unit}{Character. Unit of time: "secs", "millisecs", "mins", or "hours".
Can also be set globally via \code{options(pipetime.time_unit = "secs")}. Defaults to "secs".}
}
\value{
The input object, unchanged. Timing messages are printed or logged separately.
}
\description{
Record the runtime of a pipeline (|>) operation. Can print to the console
or log to a file. Arguments can also be set globally via options().
}
\examples{
library(dplyr)
data.frame(x = 1:3) |>
mutate(y = {Sys.sleep(0.5); x*2 }) |>
time_pipe("calc 1") |>
mutate(z = {Sys.sleep(0.5); x/2 }) |>
time_pipe("total pipeline")

}
